connector/requirements.txt
haversine==2.5.1; python_version >= "3.5" \
    --hash=sha256:6e76933a1042e6e62464b66ef686331e2798147eeedcb32b503db696d7bf0582 \
    --hash=sha256:357e41dfddc4a0f2b1c941d92a590cac840f7ce4b3da14b45b68d968b3ad7cc7
mysqlclient==1.4.6 \
    --hash=sha256:4c82187dd6ab3607150fbb1fa5ef4643118f3da122b8ba31c3149ddd9cf0cb39 \
    --hash=sha256:9e6080a7aee4cc6a06b58b59239f20f1d259c1d2fddf68ddeed242d2311c7087 \
    --hash=sha256:f646f8d17d02be0872291f258cce3813497bc7888cd4712a577fd1e719b2f213 \
    --hash=sha256:f3fdaa9a38752a3b214a6fe79d7cae3653731a53e577821f9187e67cbecb2e16
paho-mqtt==1.5.0 \
    --hash=sha256:e3d286198baaea195c8b3bc221941d25a3ab0e1507fc1779bdb7473806394be4
